requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.10.3
redis==3.3.6
vk-api==11.5.0
